+++ app/client/src/contexts/content.tsx
@@ -29,6 +29,7 @@ export type Content = {
   }>;
   parameters: {
     debounceMilliseconds: number;
+    selectOptionsPageSize: number;
   };
 };
