+++ app/client/src/routes/apiDocs.tsx
@@ -1,7 +1,5 @@
 import SwaggerUI from 'swagger-ui-react';
 import 'swagger-ui-react/swagger-ui.css';
-// components
-import { NavBar } from 'components/navBar';
 // config
 import { serverUrl } from 'config';
 
@@ -10,7 +8,6 @@ export default ApiDocs;
 export function ApiDocs() {
   return (
     <>
-      <NavBar />
       <div>
         <SwaggerUI
           url={`${serverUrl}/api/getFile/path/swagger/attains.json`}
