+++ app/client/src/routes/home.tsx
@@ -20,7 +20,7 @@ import { InfoTooltip } from 'components/infoTooltip';
 import { Loading } from 'components/loading';
 import { DownloadModal } from 'components/downloadModal';
 import { ClearSearchModal } from 'components/clearSearchModal';
-import { NavButton } from 'components/navBar';
+import { NavButton } from 'components/navButton';
 import { RadioButtons } from 'components/radioButtons';
 import { SourceSelect } from 'components/sourceSelect';
 import { Summary } from 'components/summary';
@@ -414,23 +414,11 @@ function FilterFields({
     }),
   );
 
-  // Store each row as a tuple with its row key
-  const rows: Array<[Array<[JSX.Element, string]>, string]> = [];
-  for (let i = 0; i < fieldsJsx.length; i += 2) {
-    const row = fieldsJsx.slice(i, i + 2);
-    const rowKey = row.reduce((a, b) => a + '-' + b[1], 'row');
-    rows.push([row, rowKey]);
-  }
-
   return (
-    <div>
-      {rows.map(([row, rowKey]) => (
-        <div className="grid-gap grid-row" key={rowKey}>
-          {row.map(([field, fieldKey]) => (
-            <div className="tablet:grid-col" key={fieldKey}>
-              {field}
-            </div>
-          ))}
+    <div className="grid-gap grid-row">
+      {fieldsJsx.map(([field, key]) => (
+        <div className="desktop:grid-col-4 tablet:grid-col-6" key={key}>
+          {field}
         </div>
       ))}
     </div>
@@ -611,12 +599,12 @@ function SourceSelectFilter(
   >,
 ) {
   const { sourceLabel, sourceHandler, ...selectFilterProps } = props;
-  const { profile, sourceKey, sourceValue, staticOptions } = selectFilterProps;
+  const { sourceKey, sourceValue, staticOptions } = selectFilterProps;
 
   return (
     <SourceSelect
       label={sourceLabel}
-      sources={getOptions(profile, sourceKey, staticOptions)}
+      sources={getStaticOptions(sourceKey, staticOptions)}
       onChange={sourceHandler}
       selected={sourceValue}
     >
@@ -652,10 +640,12 @@ function SelectFilter<
       profile,
       fieldName: filterKey,
       direction: sortDirection,
+      dynamicOptionLimit: content.data?.parameters.selectOptionsPageSize,
       secondaryFieldName: secondaryFilterKey,
       staticOptions,
     });
   }, [
+    content,
     contextFilters,
     defaultOption,
     filterKey,
@@ -1150,6 +1140,7 @@ function filterDynamicOptions({
   direction = 'asc',
   fieldName,
   filters,
+  limit = 20,
   profile,
   secondaryFieldName,
   staticOptions,
@@ -1158,6 +1149,7 @@ function filterDynamicOptions({
   direction?: SortDirection;
   fieldName: string;
   filters?: FilterQueryData;
+  limit?: number;
   profile: string;
   secondaryFieldName?: string | null;
   staticOptions?: StaticOptions;
@@ -1170,7 +1162,7 @@ function filterDynamicOptions({
     const data = {
       text: inputValue,
       direction: direction ?? null,
-      limit: dynamicOptionLimit,
+      limit,
       filters,
       additionalColumns: secondaryFieldName ? [secondaryFieldName] : [],
     };
@@ -1199,6 +1191,7 @@ function filterDynamicOptions({
 // Filters options by search input, returning a maximum number of options
 function filterOptions({
   defaultOption,
+  dynamicOptionLimit,
   fieldName,
   filters = {},
   profile,
@@ -1207,6 +1200,7 @@ function filterOptions({
   secondaryFieldName,
 }: {
   defaultOption?: Option | null;
+  dynamicOptionLimit?: number;
   fieldName: string;
   filters?: FilterQueryData;
   profile: string;
@@ -1225,6 +1219,7 @@ function filterOptions({
       direction,
       fieldName,
       filters,
+      limit: dynamicOptionLimit,
       profile,
       secondaryFieldName,
       staticOptions,
@@ -1373,20 +1368,6 @@ function getMultiOptionFields(fields: typeof allFieldsConfig) {
       return field.type === 'multiselect' ? field.key : null;
     }),
   );
-}
-
-// Retrieves all possible options for a given field
-function getOptions(
-  profile: string,
-  field: string,
-  staticOptions: StaticOptions,
-) {
-  const options = getStaticOptions(field, staticOptions);
-  if (options !== null) {
-    return Promise.resolve(options);
-  } else {
-    return filterDynamicOptions({ profile, fieldName: field })('');
-  }
 }
 
 function getPageName() {
@@ -1709,7 +1690,6 @@ function storageAvailable(
 ## Constants
 */
 
-const dynamicOptionLimit = 20;
 const staticOptionLimit = 100;
 
 const {
