+++ app/server/app/content/config/parameters.json
@@ -1,3 +1,4 @@
 {
-  "debounceMilliseconds": 250
+  "debounceMilliseconds": 250,
+  "selectOptionsPageSize": 20
 }
